refactored package structure of data_view
