--- src/org/uem/dam/GestorFarmacia/view/submenus/data_view/data_panel/InspectorDataPanel.java
@@ -1,9 +1,0 @@
-package org.uem.dam.GestorFarmacia.view.submenus.data_view.data_panel;
-
-import org.uem.dam.GestorFarmacia.model.DBItem;
-
-public interface InspectorDataPanel<T extends DBItem> {
-
-	public void refreshData(T dataItem);
-
-}
