+++ src/org/uem/dam/GestorFarmacia/view/submenus/data_view/list_perspective/ItemListContainer.java
@@ -1,4 +1,4 @@
-package org.uem.dam.GestorFarmacia.view.submenus.data_view;
+package org.uem.dam.GestorFarmacia.view.submenus.data_view.list_perspective;
 
 import java.util.ArrayList;
 
@@ -11,10 +11,11 @@ import org.uem.dam.GestorFarmacia.control.listener.ItemListContnListener;
 import org.uem.dam.GestorFarmacia.model.DBItem;
 import org.uem.dam.GestorFarmacia.model.RootItem;
 import org.uem.dam.GestorFarmacia.view.submenus.DefaultSubmenu;
+import org.uem.dam.GestorFarmacia.view.submenus.data_view.UpdatableDataContainer;
 
 import net.miginfocom.swing.MigLayout;
 
-public class ItemListContainer extends DefaultSubmenu<ItemListContnListener> implements DataContainer {
+public class ItemListContainer extends DefaultSubmenu<ItemListContnListener> implements UpdatableDataContainer {
 
 	private static final long serialVersionUID = 1L;
